requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-filter==25.1
dj-database-url==2.2.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
whitenoise==6.9.0
gunicorn==22.0.0
